setup-proxy-user: Warn about sshd without AllowTcpForwarding no
This is worth warning about because it's probably the best hole in the
system, letting users connect to pretty much anything else they like
from the remote host, and I'd imagine that would be something you don't
want in a lot of use cases.

However, it's not a very good check; it just looks for an uncommented
`AllowTcpForwarding no` in `/etc/ssh/sshd_config`. Not an unreasonable
heuristic, and I can't think of anything better given how complex this
configuration can be.