setup-proxy-user: Remove the `proxy` diff as was too noisy
Since in any reasonable configuration there are invariably changes from
the template provided, we were always getting output, and it was too
noisy for the information it provided, in particular starting to obscure
other warning messages.

`$HOME/proxy` is simple enough in almost all cases that the sysadmin can
just look at it to see if it looks reasonable.